Break apart progress components
