+++ src/app/progress/Milestone.tsx
@@ -1,43 +1,32 @@
 import {
-  DestinyActivityDefinition,
-  DestinyChallengeStatus,
-  DestinyDisplayPropertiesDefinition,
   DestinyMilestone,
-  DestinyMilestoneDefinition,
-  DestinyMilestoneQuest,
-  DestinyMilestoneRewardCategoryDefinition,
-  DestinyMilestoneRewardEntry,
-  DestinyObjectiveProgress,
-  DestinyQuestStatus,
-  DestinyActivityModifierDefinition,
   DestinyCharacterComponent
-  } from 'bungie-api-ts/destiny2';
-import classNames from 'classnames';
-import { t } from 'i18next';
+} from 'bungie-api-ts/destiny2';
 import * as React from 'react';
-import * as _ from 'underscore';
 import { D2ManifestDefinitions } from '../destiny2/d2-definitions.service';
 import BungieImage from '../dim-ui/BungieImage';
-import { settings } from '../settings/settings';
 import { $state } from '../ngimport-more';
 import './milestone.scss';
-
-interface MilestoneProps {
-  milestone: DestinyMilestone;
-  defs: D2ManifestDefinitions;
-  character: DestinyCharacterComponent;
-}
+import RewardActivity from './RewardActivity';
+import AvailableQuest from './AvailableQuest';
 
 /**
  * A Milestone is an activity or event that a player can complete to earn rewards.
  * There are several forms of Milestone.
  */
-export function Milestone(props: MilestoneProps) {
-  const { milestone, defs, character } = props;
-
+export function Milestone({
+  milestone,
+  defs,
+  character
+}: {
+  milestone: DestinyMilestone;
+  defs: D2ManifestDefinitions;
+  character: DestinyCharacterComponent;
+}) {
   const milestoneDef = defs.Milestone.get(milestone.milestoneHash);
 
   if (milestone.availableQuests) {
+    // A regular milestone
     return (
       <>
         {milestone.availableQuests.map((availableQuest) =>
@@ -51,6 +40,7 @@ export function Milestone(props: MilestoneProps) {
       </>
     );
   } else if (milestone.vendors) {
+    // A vendor milestone (Xur)
     const click = () => $state.go('destiny2.vendor', { id: milestone.vendors[0].vendorHash, characterId: character.characterId });
 
     return (
@@ -70,6 +60,7 @@ export function Milestone(props: MilestoneProps) {
       </div>
     );
   } else if (milestone.rewards) {
+    // Special account-wide milestones (clan)
     const rewards = milestone.rewards[0];
     const milestoneRewardDef = milestoneDef.rewards[rewards.rewardCategoryHash];
 
@@ -90,225 +81,4 @@ export function Milestone(props: MilestoneProps) {
   }
 
   return null;
-}
-
-interface RewardActivityProps {
-  rewardEntry: DestinyMilestoneRewardEntry;
-  milestoneRewardDef: DestinyMilestoneRewardCategoryDefinition;
-}
-
-/**
- * For profile-wide milestones with rewards, these show the status of each reward. So
- * far this is only used for the "Clan Objectives" milestone.
- */
-function RewardActivity(props: RewardActivityProps) {
-  const { rewardEntry, milestoneRewardDef } = props;
-
-  const rewardDef = milestoneRewardDef.rewardEntries[rewardEntry.rewardEntryHash];
-
-  const checkClass = (rewardEntry.redeemed ? 'fa-check-circle' : rewardEntry.earned ? 'fa-check-circle-o' : 'fa-circle-o');
-  const tooltip = (rewardEntry.redeemed ? 'Progress.RewardRedeemed' : rewardEntry.earned ? 'Progress.RewardEarned' : 'Progress.RewardNotEarned');
-
-  return (
-    <div className={classNames('milestone-reward-activity', { complete: rewardEntry.earned })} title={t(tooltip)}>
-      <i className={classNames('fa', checkClass)}/>
-      <BungieImage src={rewardDef.displayProperties.icon} />
-      <span>{rewardDef.displayProperties.name}</span>
-    </div>
-  );
-}
-
-interface AvailableQuestProps {
-  defs: D2ManifestDefinitions;
-  milestoneDef: DestinyMilestoneDefinition;
-  availableQuest: DestinyMilestoneQuest;
-}
-
-/**
- * Most milestones are represented as a quest, with some objectives and a reward associated with them.
- */
-function AvailableQuest(props: AvailableQuestProps) {
-  const { defs, milestoneDef, availableQuest } = props;
-
-  const questDef = milestoneDef.quests[availableQuest.questItemHash];
-  const displayProperties: DestinyDisplayPropertiesDefinition = questDef.displayProperties || milestoneDef.displayProperties;
-
-  let activityDef: DestinyActivityDefinition | null = null;
-  let modifiers: DestinyActivityModifierDefinition[] = [];
-  if (availableQuest.activity) {
-    activityDef = defs.Activity.get(availableQuest.activity.activityHash);
-    if (availableQuest.activity.modifierHashes) {
-      modifiers = availableQuest.activity.modifierHashes.map((h) => defs.ActivityModifier.get(h));
-    }
-  }
-
-  // Only look at the first reward, the rest are screwy (old engram versions, etc)
-  const questRewards = questDef.questRewards ? _.take(questDef.questRewards.items, 1).map((r) => defs.InventoryItem.get(r.itemHash)) : [];
-
-  const objectives = availableQuest.status.stepObjectives;
-  const objective = objectives.length ? objectives[0] : null;
-  const objectiveDef = objective ? defs.Objective.get(objective.objectiveHash) : null;
-
-  const tooltip = availableQuest.status.completed ? 'Progress.RewardEarned' : 'Progress.RewardNotEarned';
-
-  return (
-    <div className="milestone-quest">
-      <div className="milestone-icon" title={t(tooltip)}>
-        <BungieImage src={displayProperties.icon} />
-        <MilestoneObjectiveStatus objective={objective} status={availableQuest.status} defs={defs} />
-      </div>
-      <div className="milestone-info">
-        <span className="milestone-name">{displayProperties.name}</span>
-        {activityDef && activityDef.displayProperties.name !== displayProperties.name &&
-          <div className="milestone-location">{activityDef.displayProperties.name}</div>}
-        <div className="milestone-description">{objectiveDef ? objectiveDef.progressDescription : displayProperties.description}</div>
-        {modifiers.map((modifier) =>
-          <ActivityModifier key={modifier.hash} modifier={modifier}/>
-        )}
-        {questRewards.map((questReward) =>
-          <div className="milestone-reward" key={questReward.hash}>
-            <BungieImage src={questReward.displayProperties.icon} />
-            <span>{questReward.displayProperties.name}</span>
-          </div>
-        )}
-        <Challenges defs={defs} availableQuest={availableQuest} />
-      </div>
-    </div>
-  );
-}
-
-interface ChallengesProps {
-  defs: D2ManifestDefinitions;
-  availableQuest: DestinyMilestoneQuest;
-}
-
-/**
- * If this quest has associated challenges, display them.
- * There doesn't seem to be any consistency about which quests do and don't have challenges, though.
- */
-function Challenges(props: ChallengesProps) {
-  const { defs, availableQuest } = props;
-
-  if (!availableQuest.challenges) {
-    return null;
-  }
-
-  // If we can, filter challenges down to the current activity.
-  let filteredChallenges = availableQuest.activity ? availableQuest.challenges.filter((c) => c.objective.activityHash === availableQuest.activity.activityHash) : availableQuest.challenges;
-
-  // Sometimes none of them match the activity, though. I don't know why.
-  if (filteredChallenges.length === 0) {
-    filteredChallenges = availableQuest.challenges;
-  }
-
-  // TODO: If we don't filter, there are duplicates. The duplicates are often for the prestige-mode versions.
-  // Not sure if we want to show them since they're dups, but the completion values would be different between
-  // them, right?
-
-  // Sometimes a quest can be completed by doing challenges from multiple activities. If that's the case, group
-  // them by activity and give each a header to help them make sense.
-  const challengesByActivity = _.groupBy(filteredChallenges, (c) => c.objective.activityHash);
-  return (
-    <>
-      {_.map(challengesByActivity, (challengeStatuses, activityHash) => {
-        const activityDef = defs.Activity.get(parseInt(activityHash, 10));
-
-        return (
-          <div key={activityHash} className="milestone-challenges">
-            {_.size(challengesByActivity) > 1 &&
-              <div className="milestone-challenges-activity-name">{activityDef.displayProperties.name}</div>
-            }
-            {challengeStatuses.map((challenge) =>
-              <Challenge key={challenge.objective.objectiveHash} defs={defs} challenge={challenge} />
-            )}
-          </div>
-        );
-      })}
-    </>
-  );
-}
-
-interface ChallengeProps {
-  defs: D2ManifestDefinitions;
-  challenge: DestinyChallengeStatus;
-}
-
-/**
- * A single challenge. A lot like an objective, but we display it closer to how it appears in-game.
- */
-function Challenge(props: ChallengeProps) {
-  const { defs, challenge } = props;
-
-  const objectiveDef = defs.Objective.get(challenge.objective.objectiveHash);
-  const icon = challenge.objective.complete ? 'fa-check-circle' : 'fa-circle-o';
-
-  return (
-    <div
-     className={classNames('milestone-challenge', { complete: challenge.objective.complete })}
-     title={objectiveDef.displayProperties.description}
-    >
-    <i className={classNames('fa', icon)}/>
-      <div className="milestone-challenge-info">
-        <div className="milestone-header">
-          <span className="milestone-challenge-name">{objectiveDef.displayProperties.name}</span>
-          {objectiveDef.completionValue > 1 &&
-            <span className="milestone-challenge-progress">{challenge.objective.progress || 0}/{objectiveDef.completionValue}</span>
-          }
-        </div>
-        <div className="milestone-challenge-description">{objectiveDef.displayProperties.description}</div>
-      </div>
-    </div>
-  );
-}
-
-interface MilestoneObjectiveStatusProps {
-  objective: DestinyObjectiveProgress | null;
-  status: DestinyQuestStatus;
-  defs: D2ManifestDefinitions;
-}
-
-/**
- * The display for a milestone quest's objective. Either a count to be shown under the icon, or a
- * checkmark if the objective has been completed but not picked up. If it's a single-step objective
- * don't display anything until it's complete, because it's obvious there's only one thing to do.
- */
-function MilestoneObjectiveStatus(props: MilestoneObjectiveStatusProps) {
-  const { objective, defs, status } = props;
-  if (objective) {
-    const objectiveDef = defs.Objective.get(objective.objectiveHash);
-
-    let progress = objective.progress || 0;
-    let completionValue = objectiveDef.completionValue;
-    if (objective.objectiveHash === 3289403948) {
-      // This is the personal clan XP progression
-      const progressDef = defs.Progression.get(540048094);
-      progress *= progressDef.steps[1].progressTotal;
-      completionValue *= progressDef.steps[1].progressTotal;
-    }
-
-    if (status.completed) {
-      return <span><i className="fa fa-check-circle-o"/></span>;
-    } else if (completionValue > 1) {
-      const formatter = new Intl.NumberFormat(settings.language);
-      return <span>{formatter.format(progress)}<wbr/>/<wbr/>{formatter.format(completionValue)}</span>;
-    }
-  }
-
-  return null;
-}
-
-function ActivityModifier(props: {
-  modifier: DestinyActivityModifierDefinition;
-}) {
-  const { modifier } = props;
-
-  return (
-    <div className="milestone-modifier">
-      <BungieImage src={modifier.displayProperties.icon}/>
-      <div className="milestone-modifier-info">
-        <div className="milestone-modifier-name">{modifier.displayProperties.name}</div>
-        <div className="milestone-modifier-description">{modifier.displayProperties.description}</div>
-      </div>
-    </div>
-  );
 }
