+++ src/app/progress/Progress.tsx
@@ -5,9 +5,8 @@ import {
   DestinyFactionProgression,
   DestinyItemComponent,
   DestinyMilestone,
-  DestinyObjectiveProgress,
-  DestinyCharacterProgressionComponent
-  } from 'bungie-api-ts/destiny2';
+  DestinyObjectiveProgress
+} from 'bungie-api-ts/destiny2';
 import { t } from 'i18next';
 import * as React from 'react';
 import {
@@ -25,10 +24,8 @@ import { Milestone } from './Milestone';
 import './progress.scss';
 import { ProgressProfile, reloadProgress, getProgressStream } from './progress.service';
 import Quest from './Quest';
-import { isWellRested } from '../inventory/store/well-rested';
-import { D2ManifestDefinitions } from '../destiny2/d2-definitions.service';
-import BungieImage from '../dim-ui/BungieImage';
 import { settings, CharacterOrder } from '../settings/settings';
+import WellRestedPerkIcon from './WellRestedPerkIcon';
 
 /* Label isn't used, but it helps us understand what each one is */
 const progressionMeta = {
@@ -381,30 +378,4 @@ export function sortCharacters(characters: DestinyCharacterComponent[], order: C
   } else {
     return characters;
   }
-}
-
-function WellRestedPerkIcon(props: {
-  defs: D2ManifestDefinitions;
-  progressions: DestinyCharacterProgressionComponent;
-}) {
-  const { defs, progressions } = props;
-  const wellRestedInfo = isWellRested(defs, progressions);
-
-  if (!wellRestedInfo.wellRested) {
-    return null;
-  }
-  const formatter = new Intl.NumberFormat(window.navigator.language);
-  const perkDef = defs.SandboxPerk.get(1519921522);
-  return (
-    <div className="well-rested milestone-quest">
-      <div className="milestone-icon">
-        <BungieImage className="perk" src={perkDef.displayProperties.icon} title={perkDef.displayProperties.description} />
-        <span>{formatter.format(wellRestedInfo.progress!)}<wbr/>/<wbr/>{formatter.format(wellRestedInfo.requiredXP!)}</span>
-      </div>
-      <div className="milestone-info">
-        <span className="milestone-name">{perkDef.displayProperties.name}</span>
-        <div className="milestone-description">{perkDef.displayProperties.description}</div>
-      </div>
-    </div>
-  );
 }
